requirements.txt
websockets==11.0.3
python-telegram-bot==20.3
pytz==2023.3
aiohttp==3.8.5